requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-browser-reload==1.11.0
django-cors-headers==4.4.0
django-jazzmin==2.6.0
django-paypal==2.0
django-taggit==4.0.0
idna==3.4
pillow==10.4.0
psycopg==3.1.10
psycopg2-binary==2.8.6
pydub==0.25.1
pyscopg2==66.0.2
python-decouple==3.8
pytube==15.0.0
pytz==2023.3
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
